requirements.txt
asgiref==3.7.2
contourpy==1.2.0
cycler==0.12.1
Django==4.2.7
django-cors-headers==4.3.0
django-rest-passwordreset==1.3.0
django-versatileimagefield==3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-extra-fields==3.7.0
filetype==1.2.0
fonttools==4.44.0
kiwisolver==1.4.5
matplotlib==3.8.1
numpy==1.26.2
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-magic==0.4.27
python-magic-bin==0.4.14
pytz==2023.3.post1
rest-framework-simplejwt==0.0.2
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
